ios: change exec to support multipart messages
Adding a new CDVType: MultiPart message type, which has an array of
messages, each of which must be "massaged" the same way as single
messages are (e.g. deserialize ArrayBuffers).

Right now, multipart messages are delivered as a single array argument
to the callback, because exec callbacks do not support multiple js
arugments at the moment.  This would be a good thing to change in the
future.